// src/types/data.tsx
export interface IPageProps {
    title: string;
}

export interface ISlideImage {
    image: string;
}

export interface IScreenWidthData {
    width: number;
    slidesPerView: number;
}

export interface ElementWidthConfig {
    screenWidth: number;
    elementWidth: number;
}

export interface IWidthElement {
    screenWidth: number;
    elementWidth: number;
}
  